refactor: create redux connect container for RoomsPage
